% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{getBED}
\alias{getBED}
\title{getBED}
\usage{
getBED(x, chromosomes = "UCSC")
}
\arguments{
\item{x}{name of dataset to export}

\item{chromosomes}{style to convert chromosomes names to, default to UCSC, leave blank for WB/Ensembl}
}
\value{
BED file
}
\description{
getBED
}
